+++ packages/linkfive_purchases/lib/logic/upgrade_downgrade_purchases.dart
@@ -11,7 +11,7 @@ extension UpgradeDowngradePurchases on LinkFivePurchasesMain {
   ///   Check https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams.ProrationMode for more information
   ///
   Future<bool> handleAndroidSwitchPlan(
-      LinkFiveVerifiedReceipt oldPurchaseDetails, ProductDetails productDetails,
+      LinkFiveVerifiedReceipt oldPurchaseDetails, LinkFiveProductDetails productDetails,
       {ProrationMode? prorationMode}) async {
     GooglePlayPurchaseDetails? _oldPurchase =
         await getAndroidPurchase(oldPurchaseDetails);
@@ -24,7 +24,7 @@ extension UpgradeDowngradePurchases on LinkFivePurchasesMain {
 
     // create the upgrade or downgrade Purchase param
     GooglePlayPurchaseParam googlePlayPurchaseParam = GooglePlayPurchaseParam(
-        productDetails: productDetails,
+        productDetails: productDetails.productDetails,
         changeSubscriptionParam: ChangeSubscriptionParam(
             oldPurchaseDetails: _oldPurchase, prorationMode: prorationMode));
 
@@ -53,9 +53,9 @@ extension UpgradeDowngradePurchases on LinkFivePurchasesMain {
 
   /// Handles the Up and Downgrade of a Subscription plans
   /// Apple does not need the existing plan.
-  Future<bool> handleIosSwitchPlan(ProductDetails productDetails) async {
+  Future<bool> handleIosSwitchPlan(LinkFiveProductDetails productDetails) async {
     // init the purchase
     return InAppPurchase.instance.buyNonConsumable(
-        purchaseParam: PurchaseParam(productDetails: productDetails));
+        purchaseParam: PurchaseParam(productDetails: productDetails.productDetails));
   }
 }