+++ packages/linkfive_purchases/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.5.0
+Breaking change: fetchSubscriptions returns a Future<LinkFiveSubscriptionData?> instead of Future<List<ProductDetails>?>
+> LinkFiveSubscriptionData has `List<LinkFiveProductDetails> linkFiveSkuData` which contains the productDetail.
+
+
+
 ## 1.4.0
 Add Switch Plan functionality. You can now call LinkFivePurchases.switchPlan(...) to switch from one plan to another.
 Add more documentation